Create a new view for State objects that handles all default RESTFul API actions

+++ api/v1/views/states.py
@@ -6,17 +6,6 @@ from models import storage
 from models.state import State
 
 
-def get_state(state_id):
-    """get a state by id"""
-    state = storage.get(State, state_id)
-    return state if state else abort(404)
-
-
-def json_err(msg, stat_code):
-    """get json err"""
-    return jsonify({msg}), stat_code
-
-
 @app_views.route('/states', methods=['GET'], strict_slashes=False)
 def list_all_states():
     """List all states"""
@@ -27,17 +16,21 @@ def list_all_states():
 @app_views.route('/states/<state_id>', methods=['GET'], strict_slashes=False)
 def retrieve_a_state(state_id):
     """Retrieve a state"""
-    return jsonify(get_state(state_id).to_dict())
+    state = storage.get(State, state_id)
+    return jsonify(state.to_dict() if state else abort(404))
 
 
 @app_views.route('/states/<state_id>', methods=['DELETE'],
                  strict_slashes=False)
 def delete_a_state(state_id):
     """Delete a state"""
-    state = get_state(state_id)
-    storage.delete(state)
-    storage.save()
-    return jsonify({}), 200
+    state = storage.get(State, state_id)
+    if state:
+        storage.delete(state)
+        storage.save()
+        return jsonify({}), 200
+    else:
+        abort(404)
 
 
 @app_views.route('/states', methods=['POST'], strict_slashes=False)
@@ -45,9 +38,9 @@ def create_a_state():
     """Create a state"""
     data = request.get_json()
     if not data:
-        return json_err("Not a JSON", 400)
+        return jsonify({"Not a JSON"}), 400
     if 'name' not in data:
-        return json_err("Missing name", 400)
+        return jsonify({"Missing name"})
     new = State(**data)
     new.save()
     return jsonify(new.to_dict()), 201
@@ -56,10 +49,12 @@ def create_a_state():
 @app_views.route('/states/<state_id>', methods=['PUT'], strict_slashes=False)
 def update_a_state(state_id):
     """Update a state"""
-    state = get_state(state_id)
+    state = storage.get(State, state_id)
     data = request.get_json()
+    if not state:
+        abort(404)
     if not data:
-        return json_err("Not a JSON", 400)
+        return jsonify({"Not a JSON"}), 400
     for key, val in data.items():
         if key not in ['id', 'created_at', 'updated_at']:
             setattr(state, key, val)
